[Webserver] Increase RAM and add stack symbols to linker script
The current webserver application needs at least 16 MiB of RAM to fit.  Either
the QEMU model needs to be updated to have more RAM than the actual device or
the webserver applicaition needs to be modified to run from Flash.

This also adds stack symbols that the RISC-V port expects