Two fixes for the stack policy.
1) By default, FreeRTOS reuses main's stack as a stack for interrupt handlers
once the scheduler has started and main won't be returned to.  This causes a
problem for the stack policy, since the metadata for main's stack frame is still
there.  Happily FreeRTOS has an option configISR_STACK_SIZE_WORDS that uses a
different stack for interrupts.

2) Unhappily, the configISR_STACK_SIZE_WORDS option doesn't seem to have been
tested.  When using the suggested size of 500 words, we get a misaligned stack
that trips up an assert later on.  There was some math that tried to align the
stack, but it was wrong.  I changed it to math that definitely aligns the stack,
but is probably overly conservative.