Libxil: Moved libxil sources to webserver demo directory
This moves the source that was in RISCV_HiFive1_GCC/bsp/libxil to
FreeRTOS_Plus_FAT_TCP_CGI_RISC-V/libxil. Since the requirement of
libxil for a risc-v platform is fairly unique to the platform running
the webserver demo application, it makes sense to move the libxil
source into the only demo application that will be using it.

This commit does not include makefile integration

Change-Id: I475aa4b460bf38b24eaa318acd32efc37f92d333