ISP Test Support: Added Make target for building variants of libfreertos
This allows a test to build a variant of libfreertos with a different
name. This is useful in combination with the ability to pass in cflags
as well as test specific FreeRTOSConfig.h files. This would allow
tests that require preemption support to have a different libfreertos
than those that require the absence of that feature.

Change-Id: Ib5da571fee1175b678a249c39aba5d409c06f80f