Apply upstream AXI DMA bug fix and fix assertion
This bug fix is from [1]; the completed status of cyclic BDs was not cleared in
XAxiDma_BdRingFromHw. This currently works because we only advance to the next
BD when taking an interrupt, but the next commit will change this and end up
getting caught in an infinite packet processing loop once the entire BD ring
had been used once.

Now that XAxiDma_BdRingFromHw is clearing the completed status bit for
cyclic DMA, we cannot assert that it is present and have to just trust
that the library has not done stupid things.

[1] https://github.com/Xilinx/embeddedsw/commit/c013e82ce4817b09038e942a3b84d3320608bc78